Fix page creation without @INPUT@ placeholder
When a plain `{{NEWPAGE}}` or {{NEWPAGE>ns?nohide}} is used, the form
did not create the new page as expected, because the javascript submit
event looked for an @INPUT@ placeholder that was not there, resulting in
the user's input not being added to the new page's name.

We now detect whether the placeholder is present, and if not append
the user's input to the namespace to generate the new page id.

Fixes #110